Quiet the style gate, and close what a coverage run found open
Four things, all of them started from one -O0 --coverage run of the suite.

clang-tidy on CI named four sites. getTreeBase multiplied a DomainId by a
uint32_t count and used the result as a pointer offset, so the product is now
taken in size_t, the way the byte-size computation above it already was.
The other three are includes: slot_codegen_probe needed cme/shared.hpp for
CoherencyMode rather than core/types.hpp, test_format_race needed
util/coherency.hpp for coherency::get, and test_resource_lifecycle used
neither peer.hpp nor types.hpp.

slot_codegen asks what an optimizing compiler emits for a 64 B slot transfer.
At -O0 the compiler spills every value and reloads it 8 B at a time, and
--coverage puts a counter between the halves, so no wide transfer is available
for the check to find and a FAIL there names the build rather than the code.
check_slot_codegen.sh now reads CMAKE_BUILD_TYPE and CME_COVERAGE out of the
build tree and exits 77 when the build is unoptimized. It skips in build-cov
and still passes in the Release trees.

Three failpoint checks armed a boundary and asserted the outcome without ever
naming the boundary, unlike their siblings in the same files. They print it
now, which is what nameOf is for: a failure that names the boundary is one a
reader can re-run.

Two gaps had no assertion behind them. A dax URI's @offset suffix is optional
and means offset 0, so memory_reject now opens one without a suffix and one
whose '@' carries nothing, and asserts each is refused with a BackendError
carrying ENOENT rather than an InvalidArgumentError -- the type of the refusal
is what says the URI was well formed and the device was absent. And no case
watched a RequestAgg group whose aggregator departs while a live member
remains, so agg_recovery now checks that group1 settles on that member. Which
of the two writers puts it there is not visible in the record: leave() hands
the duty on, and every member's poll tick volunteers itself for a duty it
reads as vacant. The check states the outcome and says so.

+++ src/core/policy/successor_peterson_layout.hpp
@@ -158,7 +158,8 @@ struct ClimbStep_t
     {
         return nullptr;
     }
-    return reinterpret_cast<PetersonNode_t*>(areaBase) + domainId * getNodesPerDomain(maxNumPeers);
+    return reinterpret_cast<PetersonNode_t*>(areaBase) +
+           static_cast<std::size_t>(domainId) * getNodesPerDomain(maxNumPeers);
 }
 
 }  // namespace cme::peterson_tree
